Create 1766-Tree of Coprimes
O(N^2) or O(N^2 * logM) solution: 
Find parent node for any node first. For any node, if its parent is not coprime with it, go to parent's parent until node 0.

O(NM) solution:
Since nums[i] <= 50, so we can calculate coprimes[i][j] as a matrix first.
For any node, having ancestors (size of max(nums)), we just need to iterate all possible ancestors if it exists and find out the one that is coprimes with current node and is closest to current node. Then update ancestors, and DFS to find the answer for its children.