Create 1761-Minimum Degree of a Connected Trio in a Graph
O(VE) solution:
Say f(node) denotes the number of children for node(equal to the number of edges connected to node), then degree(a, b, c) = f(a) + f(b) + f(c) - 6 if (a, b, c) is a trio. Since we have all edges, so we just iterate all edges (a, b), and for any child c of a (or b), if it is also b's child, then (a, b, c) is a trio. Traverse all trios then we will get the answer. 
To accelerate, notice that, for any edge (a, b), minimum degree for trio having edge(a, b) = min(degree(a, b, c) for any possible c) = min(f(a) + f(b) + f(c) - 6 for any possible c) = f(a) + f(b) - 6 + min(f(c) for any possible c). So if we sort nodes by the number of its children, then we just iterate all sorted nodes, once we find node c which is a common child of a and b, then trio(a, b, c) is the one with minimum degree havind edge (a, b).